requirements.txt
asgiref==3.8.1
Django==5.1.1
django-bootstrap5==24.3
django-pygments-renderer==0.0.1
Faker==30.8.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
